Update producer.py
Changed the method to publish the message to the queue that enables other services  to consume and act on this information. This ensures that the queue was available, the message was formatted correctly, and that the queueing system was able to process the message and store it for future consumption.